% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.r
\docType{data}
\name{smoking}
\alias{smoking}
\title{Effects of cigarette smoke on the human airway epithelial cell transcriptome}
\format{
A data frame with 51 rows and 22283 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE994}
}
\usage{
smoking
}
\description{
A dataset containing 22283 transcript measurements for 51 subjects
}
\keyword{datasets}
